Added updated script.js file to project. See extended desc.
Added script.js file which was updated with the following. 
Added a for loop that allows a user to add multiple employees to the page, provided they input the employees first and last name, as well as their salary.
Added a function to log the average salary and total number of employees to the console. 
Added a function to randomly select one employee and log their name to the console.